Fixtures refactor
Move all the fixture methods to the `helpers` module.
